+++ packages/commercio_ui/CHANGELOG.md
@@ -3,6 +3,10 @@
 * **[BREAKING]** Remove `mnemonic` attribute and `hasMnemonic` getter from `StatefulCommercioAccount`
 * **[BREAKING]** Make `mnemonic` argument in `storeMnemonic()` as `@required`
 * **[BREAKING]** `restoreWallet()` now returns a `WalletWithMnemonic` entity
+* **[BREAKING]** In `StatelessCommercioAccount` and `StatefulCommercioAccount` the method `sendTokens()` now accepts the parameter `recipientAddresses` instead of `recipientAddress` to support multiple recipients.
+* Adds `getDidDocument()` in `StatelessCommercioId` and `StatefulCommercioId`
+* Adds optional parameter `httpHelper` in `StatefulCommercioId` constructor
+* Adds `didDocument` in `HttpEndpoint` enum
 
 ## [0.1.0] - 2020-07-17